SELECT命令的db字段改为i32类型, 其他的db字段为字节数组. zset2的score解析处理.

+++ src/rdb.rs
@@ -109,7 +109,7 @@ pub(crate) fn parse(input: &mut Reader,
             RDB_OPCODE_SELECTDB => {
                 let (db, _) = input.read_length()?;
                 cmd_handler.iter().for_each(|handler|
-                    handler.handle(Command::SELECT(&SELECT { db: db as u8 })));
+                    handler.handle(Command::SELECT(&SELECT { db: db as i32 })));
             }
             RDB_OPCODE_RESIZEDB => {
                 let (db, _) = input.read_length()?;
